WIP: CollapsingToolbar. Almost finished, working fine at this point. Next change: variable text width from expanded to collapsed state.
