--- app/src/main/java/com/myapplications/mywatchlist/ui/details/management/states/toolbar/DynamicOffsetScrollFlagState.kt
@@ -1,10 +1,0 @@
-package com.myapplications.mywatchlist.ui.details.management.states.toolbar
-
-abstract class DynamicOffsetScrollFlagState(
-    heightRange: IntRange,
-    scrollValue: Int
-) : ScrollFlagState(heightRange, scrollValue) {
-
-    protected abstract var scrollOffset: Float
-
-}
