--- app/src/main/java/com/myapplications/mywatchlist/ui/details/management/states/toolbar/ToolbarState.kt
@@ -1,11 +1,0 @@
-package com.myapplications.mywatchlist.ui.details.management.states.toolbar
-
-import androidx.compose.runtime.Stable
-
-@Stable
-interface ToolbarState {
-    val offset: Float
-    val height: Float
-    val progress: Float
-    var scrollValue: Int
-}
