--- app/src/main/res/drawable/logo_costa.xml
@@ -1,21 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="116.57dp"
-    android:height="26.8dp"
-    android:viewportWidth="116.57"
-    android:viewportHeight="26.8">
-  <path
-      android:fillColor="#FF000000"
-      android:pathData="M19.13,19.73a1.36,1.36 0,0 1,0.44 0.08,1.16 1.16,0 0,1 0.41,0.28l2.41,2.54a10,10 0,0 1,-4 3.12,13.92 13.92,0 0,1 -5.64,1.05 13.54,13.54 0,0 1,-5.35 -1,11.7 11.7,0 0,1 -4,-2.81 12.41,12.41 0,0 1,-2.53 -4.25A15.66,15.66 0,0 1,0 13.39,14.13 14.13,0 0,1 1,8 12.44,12.44 0,0 1,8.15 1a15.12,15.12 0,0 1,5.57 -1,14.5 14.5,0 0,1 2.78,0.26A14.77,14.77 0,0 1,18.94 1,11.41 11.41,0 0,1 21,2.08a10.8,10.8 0,0 1,1.7 1.41L20.66,6.25a2.86,2.86 0,0 1,-0.46 0.46,1.29 1.29,0 0,1 -0.76,0.2 1.4,1.4 0,0 1,-0.61 -0.14,6.66 6.66,0 0,1 -0.61,-0.35L17.5,6a5.51,5.51 0,0 0,-0.91 -0.45,7.09 7.09,0 0,0 -1.24,-0.35A8.21,8.21 0,0 0,13.68 5a7.6,7.6 0,0 0,-3 0.59A6.59,6.59 0,0 0,8.31 7.28a7.64,7.64 0,0 0,-1.53 2.64,10.23 10.23,0 0,0 -0.55,3.47A10.64,10.64 0,0 0,6.78 17a7.68,7.68 0,0 0,1.49 2.64,6.26 6.26,0 0,0 2.21,1.63 6.63,6.63 0,0 0,2.7 0.56,13.92 13.92,0 0,0 1.51,-0.08A7,7 0,0 0,16 21.44a5.58,5.58 0,0 0,1.11 -0.51,7.34 7.34,0 0,0 1.07,-0.79 1.93,1.93 0,0 1,0.47 -0.29A1.2,1.2 0,0 1,19.13 19.73Z"/>
-  <path
-      android:fillColor="#FF000000"
-      android:pathData="M51.05,13.39a14.14,14.14 0,0 1,-1 5.29,12.71 12.71,0 0,1 -2.77,4.26A12.56,12.56 0,0 1,43 25.78a15.75,15.75 0,0 1,-11.15 0,12.75 12.75,0 0,1 -4.31,-2.84 12.41,12.41 0,0 1,-2.77 -4.26,14 14,0 0,1 -1,-5.29 14,14 0,0 1,1 -5.29,12.46 12.46,0 0,1 2.77,-4.25A12.7,12.7 0,0 1,31.86 1,15.53 15.53,0 0,1 43,1a12.54,12.54 0,0 1,7.06 7.08A14.11,14.11 0,0 1,51.05 13.39ZM44.82,13.39a11.56,11.56 0,0 0,-0.5 -3.5,7.48 7.48,0 0,0 -1.45,-2.64A6.15,6.15 0,0 0,40.55 5.6,7.84 7.84,0 0,0 37.44,5a8,8 0,0 0,-3.14 0.58A6.15,6.15 0,0 0,32 7.25a7.32,7.32 0,0 0,-1.45 2.64,11.25 11.25,0 0,0 -0.51,3.5 11.29,11.29 0,0 0,0.51 3.52A7.32,7.32 0,0 0,32 19.55,6.15 6.15,0 0,0 34.3,21.2a8.33,8.33 0,0 0,3.14 0.56,8.21 8.21,0 0,0 3.11,-0.56 6.15,6.15 0,0 0,2.32 -1.65,7.48 7.48,0 0,0 1.45,-2.64A11.6,11.6 0,0 0,44.82 13.39Z"/>
-  <path
-      android:fillColor="#FF000000"
-      android:pathData="M68.58,5.62a2.27,2.27 0,0 1,-0.57 0.64,1.22 1.22,0 0,1 -0.76,0.22 1.87,1.87 0,0 1,-0.89 -0.26l-1.08,-0.59a9.39,9.39 0,0 0,-1.38 -0.58,5.66 5.66,0 0,0 -1.76,-0.26 3.79,3.79 0,0 0,-2.55 0.73,2.46 2.46,0 0,0 -0.84,2 1.79,1.79 0,0 0,0.51 1.31,4.14 4.14,0 0,0 1.32,0.9 14.19,14.19 0,0 0,1.87 0.7c0.7,0.21 1.42,0.44 2.15,0.7a19.23,19.23 0,0 1,2.15 0.91,6.91 6.91,0 0,1 1.87,1.33 6.27,6.27 0,0 1,1.33 2,7.12 7.12,0 0,1 0.5,2.8 9.13,9.13 0,0 1,-0.63 3.4A7.84,7.84 0,0 1,68 24.27,8.35 8.35,0 0,1 65,26.13a11,11 0,0 1,-4 0.67,12.69 12.69,0 0,1 -2.5,-0.25 15,15 0,0 1,-2.46 -0.71,14.87 14.87,0 0,1 -2.23,-1.09 9.65,9.65 0,0 1,-1.81 -1.4l1.79,-2.85a1.56,1.56 0,0 1,0.56 -0.52,1.62 1.62,0 0,1 0.78,-0.2 2.13,2.13 0,0 1,1.09 0.34l1.24,0.76a9.81,9.81 0,0 0,1.59 0.76,6.18 6.18,0 0,0 2.13,0.34 4,4 0,0 0,2.57 -0.73,2.79 2.79,0 0,0 0.92,-2.31 2.18,2.18 0,0 0,-0.51 -1.5,3.8 3.8,0 0,0 -1.32,-1A12.54,12.54 0,0 0,61 15.82c-0.7,-0.19 -1.41,-0.4 -2.14,-0.64a15.3,15.3 0,0 1,-2.15 -0.86A6.81,6.81 0,0 1,54.89 13a6.13,6.13 0,0 1,-1.32 -2.09,8.18 8.18,0 0,1 -0.51,-3.09 7.29,7.29 0,0 1,0.6 -2.88,7.48 7.48,0 0,1 1.74,-2.5 8.64,8.64 0,0 1,2.83 -1.75A10.39,10.39 0,0 1,62.06 0a13.89,13.89 0,0 1,2.35 0.19,11.5 11.5,0 0,1 2.17,0.56 11.6,11.6 0,0 1,1.92 0.88,9 9,0 0,1 1.59,1.16Z"/>
-  <path
-      android:fillColor="#FF000000"
-      android:pathData="M92.52,0.29L92.52,5.11L85.14,5.11v21.4L79.06,26.51L79.06,5.11L71.67,5.11L71.67,0.29Z"/>
-  <path
-      android:fillColor="#FF000000"
-      android:pathData="M116.57,26.51h-4.72a2.19,2.19 0,0 1,-1.3 -0.36,1.89 1.89,0 0,1 -0.71,-0.95l-1.55,-4.57h-10L96.77,25.2a2.06,2.06 0,0 1,-0.7 0.91,2 2,0 0,1 -1.28,0.4L90,26.51L100.19,0.29h6.23ZM99.72,16.51h7.16l-2.41,-7.18c-0.15,-0.46 -0.34,-1 -0.55,-1.61s-0.41,-1.29 -0.62,-2c-0.19,0.73 -0.39,1.41 -0.59,2s-0.4,1.17 -0.58,1.62Z"/>
-</vector>
