--- app/src/main/res/drawable/logo_rica.xml
@@ -1,18 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="83.52dp"
-    android:height="26.8dp"
-    android:viewportWidth="83.52"
-    android:viewportHeight="26.8">
-  <path
-      android:fillColor="#FF000000"
-      android:pathData="M6.08,16.81v9.7L0,26.51L0,0.29L8.53,0.29a17.51,17.51 0,0 1,4.85 0.58A9.24,9.24 0,0 1,16.66 2.5,6.2 6.2,0 0,1 18.5,5a8.06,8.06 0,0 1,0.58 3.09,8.28 8.28,0 0,1 -0.34,2.42 7.38,7.38 0,0 1,-1 2.09,7.51 7.51,0 0,1 -1.66,1.72 8.41,8.41 0,0 1,-2.25 1.24,6.46 6.46,0 0,1 1.14,0.76 4.79,4.79 0,0 1,0.93 1.1l5.58,9.12L16,26.54a2.34,2.34 0,0 1,-2.2 -1.15L9.4,17.75A1.83,1.83 0,0 0,8.72 17a2.19,2.19 0,0 0,-1.09 -0.22ZM6.08,12.62L8.53,12.62a6.31,6.31 0,0 0,2.12 -0.32,3.84 3.84,0 0,0 1.43,-0.87 3.4,3.4 0,0 0,0.81 -1.3,5 5,0 0,0 0.25,-1.62A3.39,3.39 0,0 0,12 5.81a5.28,5.28 0,0 0,-3.48 -1L6.08,4.81Z"/>
-  <path
-      android:fillColor="#FF000000"
-      android:pathData="M30.6,26.51L24.48,26.51L24.48,0.29L30.6,0.29Z"/>
-  <path
-      android:fillColor="#FF000000"
-      android:pathData="M53.08,19.73a1.26,1.26 0,0 1,0.43 0.08,1.19 1.19,0 0,1 0.42,0.28l2.41,2.54a10,10 0,0 1,-4 3.12,13.92 13.92,0 0,1 -5.64,1.05 13.62,13.62 0,0 1,-5.36 -1,11.66 11.66,0 0,1 -4,-2.81 12.41,12.41 0,0 1,-2.53 -4.25A15.66,15.66 0,0 1,34 13.39,14 14,0 0,1 35,8a12.62,12.62 0,0 1,2.81 -4.24A12.71,12.71 0,0 1,42.1 1a15.11,15.11 0,0 1,5.56 -1,13.8 13.8,0 0,1 5.22,1A11.5,11.5 0,0 1,55 2.08a10.28,10.28 0,0 1,1.69 1.41l-2,2.76a2.93,2.93 0,0 1,-0.47 0.46,1.26 1.26,0 0,1 -0.75,0.2 1.4,1.4 0,0 1,-0.61 -0.14,6.75 6.75,0 0,1 -0.62,-0.35L51.45,6a5.64,5.64 0,0 0,-0.92 -0.45,6.62 6.62,0 0,0 -1.24,-0.35A8.11,8.11 0,0 0,47.63 5a7.65,7.65 0,0 0,-3 0.59,6.77 6.77,0 0,0 -2.35,1.67 8,8 0,0 0,-1.53 2.64,10.49 10.49,0 0,0 -0.54,3.47A10.91,10.91 0,0 0,40.72 17a7.86,7.86 0,0 0,1.5 2.64,6.23 6.23,0 0,0 2.2,1.63 6.63,6.63 0,0 0,2.7 0.56,14 14,0 0,0 1.52,-0.08 6.84,6.84 0,0 0,1.26 -0.28A5.66,5.66 0,0 0,51 20.93a7.92,7.92 0,0 0,1.07 -0.79,1.93 1.93,0 0,1 0.47,-0.29A1.2,1.2 0,0 1,53.08 19.73Z"/>
-  <path
-      android:fillColor="#FF000000"
-      android:pathData="M83.52,26.51L78.8,26.51a2.19,2.19 0,0 1,-1.3 -0.36,2 2,0 0,1 -0.71,-0.95l-1.55,-4.57h-10L63.72,25.2a2,2 0,0 1,-0.7 0.91,2 2,0 0,1 -1.28,0.4L57,26.51L67.14,0.29h6.23ZM66.67,16.51h7.16l-2.41,-7.18c-0.15,-0.46 -0.34,-1 -0.55,-1.61s-0.41,-1.29 -0.62,-2c-0.19,0.73 -0.39,1.41 -0.59,2s-0.4,1.17 -0.58,1.62Z"/>
-</vector>
